+++ lib/services/token_validation_service.dart
@@ -1,7 +1,6 @@
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:aliens/services/auth_service.dart';
 import 'package:dart_jsonwebtoken/dart_jsonwebtoken.dart';
-import 'package:aliens/services/apis.dart';
-
 // 토큰 유효성 검사
 
 class TokenValidationService {
@@ -12,7 +11,6 @@ class TokenValidationService {
     var refreshToken = await storage.read(key: 'refreshToken');
 
     if (accessToken == null || refreshToken == null) {
-      print('로그인 필요');
       return false;
     }
 
@@ -28,28 +26,23 @@ class TokenValidationService {
     Duration refreshTokenDiff =
         DateTime.now().difference(refreshTokenTimestamp);
 
-    print(
-        '지금 시간: ${DateTime.now()}\n엑세스 토큰 저장 시간: $accessTokenTimestamp\n 차이: $accessTokenDiff');
-    print(
-        '지금 시간: ${DateTime.now()}리프레시 토큰 저장 시간: $refreshTokenTimestamp\n 차이: $refreshTokenDiff');
-
     //refresh token 유효 기간(30일) 만료
     if (refreshTokenDiff >= const Duration(days: 30)) {
-      print("refresh token 만료");
+      // 리프레시 토큰 만료
       await _clearStorage();
       return false;
     }
     // 시간 음수 오류 처리
     else if (accessTokenDiff < Duration.zero ||
         refreshTokenDiff < Duration.zero) {
-      print('시간 음수 오류');
+      // 시간 음수 오류
       await _clearStorage();
       return false;
     }
     // access token 유효 기간(1일) 만료
     else if (accessTokenDiff >= const Duration(days: 1)) {
-      print('access token 만료');
-      if (await APIs.getAccessToken()) {
+      // 액세스 토큰 만료
+      if (await AuthService.getAccessToken()) {
         return true;
       } else {
         await _clearStorage();
