+++ lib/views/pages/setting/setting_edit_PW_done_page.dart
@@ -1,9 +1,9 @@
+import 'package:aliens/services/auth_service.dart';
 import 'package:aliens/views/components/appbar.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:easy_localization/easy_localization.dart';
-import '../../../services/apis.dart';
 import '../../components/button.dart';
 
 class SettingEditPWDonePage extends StatefulWidget {
@@ -75,7 +75,7 @@ class _SettingEditPWDonePageState extends State<SettingEditPWDonePage> {
                       final fcmToken =
                           await FirebaseMessaging.instance.getToken();
                       //로그아웃
-                      APIs.logOut(context);
+                      AuthService.logOut(context);
                     }),
               )),
             ],
